+++ EDDI/MainWindow.xaml.cs
@@ -326,6 +326,12 @@ namespace Eddi
                     setUpCompanionAppStage1("Unexpected problem\r\nPlease report this at http://github.com/CmdrMcDonald/EliteDangerousDataProvider/issues\r\n" + ex);
                 }
             }
+            else if (companionAppLogoutText.Visibility == Visibility.Visible)
+            {
+                // Logged in - handle logout
+                CompanionAppService.Instance.Logout();
+                setUpCompanionAppStage1();
+            }
         }
 
         private void setUpCompanionAppStage1(string message = null)
@@ -348,7 +354,8 @@ namespace Eddi
             companionAppCodeText.Text = "";
             companionAppCodeLabel.Visibility = Visibility.Hidden;
             companionAppCodeText.Visibility = Visibility.Hidden;
-            companionAppNextButton.Visibility = Visibility.Visible;
+            companionAppLogoutText.Visibility = Visibility.Hidden;
+            companionAppNextButton.Content = "Next";
         }
 
         private void setUpCompanionAppStage2(string message = null)
@@ -369,7 +376,8 @@ namespace Eddi
             companionAppPasswordText.Visibility = Visibility.Hidden;
             companionAppCodeLabel.Visibility = Visibility.Visible;
             companionAppCodeText.Visibility = Visibility.Visible;
-            companionAppNextButton.Visibility = Visibility.Visible;
+            companionAppLogoutText.Visibility = Visibility.Hidden;
+            companionAppNextButton.Content = "Next";
         }
 
         private void setUpCompanionAppComplete(string message = null)
@@ -391,7 +399,8 @@ namespace Eddi
             companionAppCodeText.Text = "";
             companionAppCodeLabel.Visibility = Visibility.Hidden;
             companionAppCodeText.Visibility = Visibility.Hidden;
-            companionAppNextButton.Visibility = Visibility.Hidden;
+            companionAppLogoutText.Visibility = Visibility.Visible;
+            companionAppNextButton.Content = "Log out";
         }
 
         // Handle changes to the Shipyard tab