+++ Tests/GeneratorTests.cs
@@ -31,19 +31,25 @@ namespace Tests
                             System.Reflection.MethodInfo method = entry.Value.GetMethod("get_" + variable.Key);
                             if (method != null)
                             {
-                                if (method.ReturnType == typeof(string))
+                                Type returnType = method.ReturnType;
+                                if (returnType.IsGenericType && returnType.GetGenericTypeDefinition() == typeof(Nullable<>))
+                                {
+                                    returnType = Nullable.GetUnderlyingType(returnType);
+                                }
+
+                                if (returnType == typeof(string))
                                 {
                                     Console.WriteLine("  * {TXT:EDDI " + entry.Key.ToLowerInvariant() + " " + variable.Key + "} " + variable.Value);
                                 }
-                                else if (method.ReturnType == typeof(int))
+                                else if (returnType == typeof(int))
                                 {
                                     Console.WriteLine("  * {INT:EDDI " + entry.Key.ToLowerInvariant() + " " + variable.Key + "} " + variable.Value);
                                 }
-                                else if (method.ReturnType == typeof(bool))
+                                else if (returnType == typeof(bool))
                                 {
                                     Console.WriteLine("  * {BOOL:EDDI " + entry.Key.ToLowerInvariant() + " " + variable.Key + "} " + variable.Value);
                                 }
-                                else if (method.ReturnType == typeof(decimal) || method.ReturnType == typeof(double) || method.ReturnType == typeof(long))
+                                else if (returnType == typeof(decimal) || returnType == typeof(double) || returnType == typeof(long))
                                 {
                                     Console.WriteLine("  * {DEC:EDDI " + entry.Key.ToLowerInvariant() + " " + variable.Key + "} " + variable.Value);
                                 }