+++ ChangeLog.md
@@ -2,6 +2,7 @@
   * Update documentation for material and rarity information
   * Update coriolis export for Beluga Liner
   * Help large star system databases by adding an index
+  * Send raw ship JSON directly to Coriolis for import rather than use local processing
 
 #2.0.9
   * Do not update ship configuration when data is not available from the companion API